refactor: geometry logic split into a subpackage
